python-fd: merge vmware plugin into one file
