+++ core/src/vmware/vmware_plugin/CMakeLists.txt
@@ -1,6 +1,6 @@
 #   BAREOS® - Backup Archiving REcovery Open Sourced
 #
-#   Copyright (C) 2020-2020 Bareos GmbH & Co. KG
+#   Copyright (C) 2020-2023 Bareos GmbH & Co. KG
 #
 #   This program is Free Software; you can redistribute it and/or
 #   modify it under the terms of version three of the GNU Affero General Public
@@ -31,7 +31,7 @@ set(INSTALL_LIB_DIR
 mark_as_advanced(INSTALL_LIB_DIR)
 
 install(
-  FILES BareosFdPluginVMware.py bareos-fd-vmware.py
+  FILES bareos-fd-vmware.py
   PERMISSIONS OWNER_READ OWNER_WRITE OWNER_EXECUTE GROUP_READ GROUP_EXECUTE
               WORLD_READ WORLD_EXECUTE
   DESTINATION "${INSTALL_LIB_DIR}/bareos/plugins"
